+++ packages/coin-support-canton/src/services/api/transaction.ts
@@ -13,8 +13,6 @@ import {
   ICantonPrepareExternalPartyTxnParams,
   ICantonBroadcastExternalPartyTxnParams,
   ICantonInstrument,
-  ICantonWaitForTxnCompletionParams,
-  ICantonWaitForTxnCompletionResult,
 } from './types';
 
 import { config } from '../../config';
@@ -142,8 +140,8 @@ export const broadcastTransactionToBlockchain = async (
   params: ICantonBroadcastTxnParams,
   keyDB?: IKeyValueStore,
 ): Promise<{
-  commandId: string;
-  ledgerEndOffset: number;
+  updateId: string;
+  completionOffset: number;
 }> => {
   const url = `${baseURL}/broadcast`;
   const response = await makePostRequestWithAuthTokenConfig(url, params, keyDB);
@@ -152,16 +150,6 @@ export const broadcastTransactionToBlockchain = async (
     !response.data.error,
     new Error('Server: Invalid txn hash from server'),
   );
-
-  return response.data;
-};
-
-export const waitForTxnCompletion = async (
-  params: ICantonWaitForTxnCompletionParams,
-  keyDB?: IKeyValueStore,
-): Promise<ICantonWaitForTxnCompletionResult | undefined> => {
-  const url = `${baseURL}/wait-for-completion`;
-  const response = await makePostRequestWithAuthTokenConfig(url, params, keyDB);
 
   return response.data;
 };