+++ packages/cysync-core-constants/src/i18n/types.ts
@@ -140,6 +140,7 @@ interface LangDashboard {
 
 interface LangTooltips {
   downloadCsv: string;
+  filter: string;
 }
 
 interface LangDeviceAuthentication {
@@ -578,9 +579,16 @@ interface LangHistory {
     startDate: string;
     expirationDate: string;
     operation: string;
+    buttons: {
+      addRemarks: string;
+    };
   };
   noData: { text: string; subText: string; buttonText: string };
   search: { placeholder: string; notFound: { text: string; subText: string } };
+  filter: {
+    wallets: string;
+    accounts: string;
+  };
 }
 
 interface LangOnboarding {