+++ packages/cysync-core/src/components/index.ts
@@ -21,3 +21,4 @@ export * from './LangDropdown';
 export * from './DeleteAccount';
 export * from './EnableApprovalPrompt';
 export * from './EnableMergeDelegationPrompt';
+export * from './HistoryFilter';