+++ packages/app-support-inheritance/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cypherock/app-support-inheritance
 
+## 1.0.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-interfaces@0.1.1
+  - @cypherock/cysync-utils@0.1.1
+
+## 1.0.13-pre.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-interfaces@0.1.1-pre.0
+  - @cypherock/cysync-utils@0.1.1-pre.0
+
 ## 1.0.12
 
 ### Patch Changes