+++ packages/automation-scripts/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cypherock/cysync-automation-scripts
 
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [[`f4fa6c43`](https://github.com/Cypherock/cypherock-cysync/commit/f4fa6c43570802d537b975cf262d21e63024c8e1)]:
+  - @cypherock/cysync-core-constants@0.7.3
+  - @cypherock/cysync-interfaces@0.1.1
+  - @cypherock/cysync-utils@0.1.1
+
+## 0.1.20-pre.0
+
+### Patch Changes
+
+- Updated dependencies [[`f4fa6c43`](https://github.com/Cypherock/cypherock-cysync/commit/f4fa6c43570802d537b975cf262d21e63024c8e1)]:
+  - @cypherock/cysync-core-constants@0.7.3-pre.0
+  - @cypherock/cysync-interfaces@0.1.1-pre.0
+  - @cypherock/cysync-utils@0.1.1-pre.0
+
 ## 0.1.19
 
 ### Patch Changes