+++ packages/coin-support-sia/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @cypherock/coin-support-sia
 
+## 2.0.7
+
+### Patch Changes
+
+- Updated dependencies [[`f4fa6c43`](https://github.com/Cypherock/cypherock-cysync/commit/f4fa6c43570802d537b975cf262d21e63024c8e1), [`27d1c3f5`](https://github.com/Cypherock/cypherock-cysync/commit/27d1c3f5d7d69a98faf7595aaf60a4bba9411297)]:
+  - @cypherock/coin-support-utils@0.2.1
+  - @cypherock/db-interfaces@0.2.1
+  - @cypherock/coin-support-interfaces@0.1.4
+  - @cypherock/cysync-interfaces@0.1.1
+  - @cypherock/cysync-utils@0.1.1
+
+## 2.0.7-pre.0
+
+### Patch Changes
+
+- Updated dependencies [[`f4fa6c43`](https://github.com/Cypherock/cypherock-cysync/commit/f4fa6c43570802d537b975cf262d21e63024c8e1), [`27d1c3f5`](https://github.com/Cypherock/cypherock-cysync/commit/27d1c3f5d7d69a98faf7595aaf60a4bba9411297)]:
+  - @cypherock/coin-support-utils@0.2.1-pre.0
+  - @cypherock/db-interfaces@0.2.1-pre.0
+  - @cypherock/coin-support-interfaces@0.1.4-pre.0
+  - @cypherock/cysync-interfaces@0.1.1-pre.0
+  - @cypherock/cysync-utils@0.1.1-pre.0
+
 ## 2.0.6
 
 ### Patch Changes